debugd: Add missing Errors
This is the result of an audit of D-Bus ErrorPtr usage to make sure the
error is set whenever return false is returned by a handler to avoid
CHECK failures in libbrillo.

Some cleanup work was done to constants in dmesg_tool.cc for consistent
usage of constexpr and const for char[] types.

BUG=b:279097307
TEST=CQ passes

Change-Id: I141c3903600055c9634c90f55e142b665dbdab9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4459718
Reviewed-by: Aashay Shringarpure <aashay@google.com>
Tested-by: Allen Webb <allenwebb@google.com>
Commit-Queue: Allen Webb <allenwebb@google.com>