printscanmgr: refactor upstart_tools
Move upstart_tools to inside the executor, where it will have the root
privilege it needs to send D-Bus messages to the upstart daemon. Change
the interface to use Mojo enums. In an effort to not give the executor
any more power than necessary, remove the ability to start upstart jobs
since this functionality is not used in printscanmgr.

BUG=b:257070388
TEST=FEATURES=test emerge-nami printscanmgr

Change-Id: I4cc36d28fccf4dc7c8f8b364d8c5dff2fe683b0e
Disallow-Recycled-Builds: test-failures
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4458457
Reviewed-by: Nathan Muggli <nmuggli@google.com>
Commit-Queue: Paul Moy <pmoy@chromium.org>
Tested-by: Paul Moy <pmoy@chromium.org>