hiberman: Don't try to resume if the hibernate cookie is uninitialized
Currently an uninitialized cookie causes hiberman to try a resume,
which is almost certainly going to fail. The typical case of an
uninitialized hibernate cookie would be booting after an
installation from USB.

Treat an uninitialized hibernate cookie like a NoResume cookie. If
an uninitialized cookie is found set it to NoResume.

BUG=b:279220948
TEST=on brya:
  # clear hibernate cookie
  dd if=/dev/zero bs=1 count=16 | dd of=/dev/nvme0n1 bs=1 seek=992
  reboot
    => no reboot loop

Change-Id: If885226bcafe2b2241492457f64ae23f434471ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4469334
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Brian Geffon <bgeffon@chromium.org>