secagentd: Handle mount namespaces while computing image checksum
The Process exec BPF passes down a new field `pid_for_setns` that's set
to the earliest process in the hierarchy that has the same mount
namespace. The motivation is to find a process that's more likely to be
alive by the time we hash the image.

procfs code already relies on the process being alive when it runs so
the same process' pid is used as `pid_for_setns`.

The image is then looked up in `/proc/pid_for_setns/root/` for stating
and hashing.

BUG=b:253661187
TEST=https://paste.googleplex.com/5094352626909184 is a process exec
TEST=from ARC container that was incomplete before.
TEST=secagentd.ProcessEvents tast test still passes on a VM.

Change-Id: Ifa5c3cfb433feb33f385f12bacf4ab7720f0e3d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4472568
Tested-by: Aashay Shringarpure <aashay@google.com>
Commit-Queue: Aashay Shringarpure <aashay@google.com>
Reviewed-by: Jason Ling <jasonling@chromium.org>
Reviewed-by: Ryan Borzello <rborzello@google.com>