shill: net: move IPAddress's bitwise operation to private
The bitwise operation of IPAddress (MaskWith and MergeWith) is not
used outside the class. Also, these methods will be removed when we
migrate the internal data of IPAddress from ByteBuffer to the new CIDR
classes.

This CL moves these methods from public to private, and removes the
corresponding test cases. It's part of the preparation of the CIDR
migration.

BUG=b:269983153
TEST=FEATURES=test emerge-hatch shill-net shill patchpanel

Change-Id: I6940da8c57621773694e41ad97d459d93a136c74
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4473150
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>