cryptohome: create AsyncInitPtr
There's a common pattern with some global objects which have to be
initialized on the mount thread, that we pass around callbacks to the
objects which all the uses of the object have to then call to see if
they can get a valid pointer.

To clean this up and make it a little neater, as well as making the
pattern a bit more explicit, create a utility template here that
basically this for you, packed up in a pointer interface.

BUG=b:276453357
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I90a36a7af4276e1e10bfc48b46ec3d575babf79e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4475739
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Yi Chou <yich@google.com>