login: Force Chromad devices to fail booting
Make session manager initialization fail on devices that are still in
AD management mode. This is done by intentionally crashing the service,
and rebooting the device. This can only happen to devices that update to
M115+ already in AD mode, because Chrome devices can no longer be
enrolled in this mode, since M110 (b/259180126).

With this change, it'll be safe to assume, everywhere else in the
client-side code base, that a device in M115+ is not in AD mode.
Therefore, the clean-up described in b/279364186 will be safe.

BUG=b:279364186
TEST=cros_run_unit_tests --board=${BOARD} "chromeos-login"

Change-Id: Id538c1462ef51567b482c7855cab13050f0067d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4480191
Commit-Queue: Felipe Andrade <fsandrade@chromium.org>
Reviewed-by: Igor <igorcov@chromium.org>
Reviewed-by: Roland Bock <rbock@google.com>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: Felipe Andrade <fsandrade@chromium.org>