missive: Only run minimum required CQ TAST tests.
This changes adds a DIR_METADATA file which tells CQ to only run the
test specified in the cq_minimal_each_arch test plan when submitting
changes to missive. It's ok to skip running most of the tast tests
during CQ since they are not related to missive. Once we create a test
plan with reporting-related tast tests, we can add that here as well.

BUG=b:274635390
TEST=none

Change-Id: Ia7465a3dd13276912cc07a256aa7b0c17a3c9ffd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4481274
Tested-by: Josh Hilke <jrhilke@google.com>
Commit-Queue: Josh Hilke <jrhilke@google.com>
Reviewed-by: Leonid Baraz <lbaraz@chromium.org>
Reviewed-by: Sergey Frolov <sfrolov@google.com>