common-mk: Check to ensure including source files
Occasionally, a source file, especially a unit test file, is added to
the repo but is not included in the build.  This kind of mistake can
often evade code review, or is discovered only later in the review stage
and wasted time for early discovery of errors. This CL adds a script
examines each *.cc file and checks whether they are present in a gn file
in their project directory.

BUG=b:273373137
FIXED=b:273373137
TEST=Start a new branch in platform2/. Create 3 commits: the first one
adds a file "some-directory/sub/abc_test.cc", the second one adds a file
"some-directory/a", and the third one adds a file
"some-directory/sub/added_test.cc" and adds it to a gn file. Run
presubmit hooks. The first commit shows an error that the test file is
not included in any BUILD.gn files, and the second and third commits
show no error.

Cq-Depend: chromium:4507447
Change-Id: Ib634cf19dd2a742f43e7c49a662e4dd1bd0921be
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4481277
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Hong Xu <xuhong@google.com>
Commit-Queue: Hong Xu <xuhong@google.com>