shill: net: deprecate IPAddress(Family) constructor
The behavior of shill::IPAddress(Family) constructor is weird. It
creates a instance but not initialize the internal data. In this state
IsValid() is false but IsDefault() is true. We should avoid this weird
state.

This CL changes the IPAddress(Family) constructor to a static method
CreateFromFamily_Deprecated(), and adds another CreateFromFamily()
method that initializes the returned instance. Then it updates all
the callers to CreateFromFamily_Deprecated() except these two cases:

- Use CreateFromFamily(kFamilyUnknown) to replace
  IPAddress(kFamilyUnknown), because the behavior is the same
  as the original one.

- Use CreateFromFamily(family) if the caller calls
  SetAddressToDefault() or SetAddressAndPrefixFromString() immediately
  after creation.

This CL should not change any logic.

BUG=b:279724438
TEST=FEATURES=test emerge-hatch shill-net shill

Change-Id: Ie32b84663e4ee9f0c2bcc95381c0deb9d322eb4a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4486560
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>