cryptohome: Move CreateCredentialVerifier to the driver
This required shuffling around some construction again in order to make
the necessary objects available to the AuthFactorDriver.

First, it required pulling ownership of the Legacy FP block service up
into the UDA; this wasn't that big of a deal as the UDA was already
constructing it, and it's consistent with how the new FP block service
is already being owned.

Second, I had to pass down the challenge credential objects. I chose to
solve this using AsyncInitPtr instead of reusing the existing
AuthBlockUtility strategy of having an extra init phase that it calls. I
think the async ptr makes things simpler than the multi-phase init, and
is also much easier to just change to a normal pointer if we end up
managing to simplify away the multi-thread init in UDA eventually.

BUG=b:278937647
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Ib3b81a66b95625f16a3be67956c7c77be5e7d7ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4496972
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>