crash: Update log message when ID 0000000000000001 is received
The crash report server has quotas on the number of crash reports it
will accept for a particular product/version combination. When this
quota is exceeded, it will return good status to the device (HTTP 200)
along with a special crash report receipt ID of "0000000000000001".
This prevents the device from re-attempting the crash report upload due
to bad status (i.e., HTTP 429: "Too Many Requests").

This CL updates the log message when the special ID is received, to make
it clearer to developers that the crash report was dropped due to
throttling (and will not be available at go/crash for debugging). This
should help prevent any confusion when seeing this special crash report
ID in (for example) feedback reports.

BUG=b:278914110
TEST=cros_workon_make --test --board=skyrim crash-reporter
TEST=cros_workon_make --test --board=amd64-generic crash-reporter

Change-Id: I29c872581ff79570db334c2f10a7e03f0d72cb76
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4495003
Commit-Queue: Tim Van Patten <timvp@google.com>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Tested-by: Tim Van Patten <timvp@google.com>