flex_bluetooth: Add kDisableLEGetVendorCapabilities to 04ca:3015
The BT adapter 04ca:3015 is not working on Floss: it continually
cycles off and on because it is unable to make use of the command for
LE get vendor capabilities. Add the override
kDisableLEGetVendorCapabilities to prevent Floss from sending the
command when this BT adapter is in use.

Additionally, ordered the kAdapterSyspropOverrides entries in numerical
order according to the vendor ID and product ID of the BT adapters

BUG=b:273381097
TEST=`emerge-reven flex_bluetooth chromeos-bsp-reven` and deployed them
  to a HP Probook 6445b containing this Bluetooth adapter. Turned Floss
  on and verified that Bluetooth is able to be switched on. Also
  verified that a BT headset is able to output audio and that its
  microphone also works.

Change-Id: I85d77a3141b507659c350e846775fd6e95e93db7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4496114
Reviewed-by: Patrick Thompson <ptf@google.com>
Commit-Queue: Adrian Chan <adchan@google.com>
Tested-by: Adrian Chan <adchan@google.com>