lorgnette: Add DeviceTracker for async discovery API
DeviceTracker will be responsible for keeping track of which devices are
available on the system and which devices are in use.  To start with,
implement a bare bones set of functions that fill in the previous
StartScannerDiscovery and StopScannerDiscovery stubs.  This gets the
classes and planned state flow in place, but still doesn't do any actual
device detection.  Actual device detection will be in follow-up CLs.

BUG=b:277049556
TEST=lorgnette_cli discover

Change-Id: Ie98a5cd399f23e72020e5a64c82936517c5e2f2d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4503648
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Paul Moy <pmoy@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>