cryptohome: Move the factor->block mapping to the driver
In this case I passed the list of block types up via the common driver
template arguments, rather than manually implementing the production of
a list in each driver.

Perhaps this is overly clever, but I felt like this might be a case
where we want to be able to do these type->type mapping statically, so
handling it via the template rather than at runtime seemed like a good
idea? Plus the boilerplate for efficiently constructing a list seemed
somewhat tedious to copy-paste into every factor type. However, if we
feel that this is just too messy I can rewrite it in a more vanilla way.

BUG=b:278937647
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Ic97a5f482081d684c45fe9644437c1e52ca171e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4500739
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Thomas Cedeno <thomascedeno@google.com>