init: Allow fwupd to modify dbx UEFI variable
The UEFI dbx variable is used to revoke bootloaders with security
flaws. Fwupd is used to apply updates to the dbx. Since fwupd is run in
a minijail sandbox and not as root, we need to modify the dbx variable
file so that fwupd can write to it. This requires changing the file
ownership, and also removing the immutable bit that the kernel sets by
default on most UEFI vars to prevent the user from accidentally making
bad changes (see [1]).

BUG=b:244312163,b:273781238
TEST=FEATURES=test emerge-reven chromeos-init
TEST=Boot from UEFI, check attributes and owner of dbx var:
TEST=lsattr /sys/firmware/efi/efivars/dbx-d719b2cb-3d3a-4596-a3bc-dad00e67656f
TEST=ls -l /sys/firmware/efi/efivars/dbx-d719b2cb-3d3a-4596-a3bc-dad00e67656f

Change-Id: Ie2f2cc54d7db7e0f02957091d3eebb12ce4c43ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4502761
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>