lorgnette: Refactor ScanHandler out of cli
The ScanHandler class represents a d-bus operation that gives its
progress as a stream of status signals.  This is a generally useful
concept for async lorgnette APIs, so refactor it out of a private class
into lorgnette_cli.cc into a reusable class.  There are no behavior
changes in this CL.

BUG=b:277049883
TEST=lorgnette_cli scan

Change-Id: Ib4339d5bbeafa63d22e5ee2f3ff9216be3d2c121
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4503649
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Paul Moy <pmoy@chromium.org>