lorgnette: Add discover verb to cli
This adds a new "discover" verb that uses the async discovery API to
list scanners.  The functionality will eventually be folded into the
main list and scan verbs, but this provides a way to test the new async
discovery API on hardware while keeping the existing API working in the
meantime.

BUG=b:277049883
TEST=lorgnette_cli discover

Change-Id: I2bc8312a3f2d54e6ee4fda621e50c9b6fd5e54ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4503650
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Paul Moy <pmoy@chromium.org>