cryptohome: Break up the auth factor driver common class
This refactors the auth factor driver setup to replace the "one big
class" common class with a set of smaller classes that you use multiple
inheritance with instead.

Generally we prefer to avoid multiple inheritance, but I think the model
here is simple enough and straightforward enough that it avoids most of
the issues that MI has. The top of the tree is a pure virtual interface;
the multiply-inherited classes are all simple ABCs which only implement
a specific narrow subset of the interface, and we use `final` carefully
to trigger compiler errors if you were to try to mix overlapping
implementations somehow.

In general the benefit is that this avoids the complexity of the
existing class getting bigger and bigger over time. Even if we add more
driver functions and more common code, it can be added with small ABCs
instead of one common class trying to add a dozen different unrelated
features.

More specifically this also makes it much easier to add optional common
code. When adding Prepare support it would be very useful to have a
common implementation of the functions for the drivers that don't
support Prepare, but this much cleaner if you do it with a separate
class rather than trying to do it in the common class. As a
demonstration of this I also refactored the CredentialVerifier code to
work in a similar way.

BUG=b:278937647
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I30d42e92306bf50df7e3c193d1b044b7475fd810
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4500490
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>