cryptohome: Move prepare to the factor drivers
This relocates the prepare functions from auth block utility to the auth
factor drivers. This removes the last piece of factor-specific (as
opposed to block-specific) code in auth block utility.

To make things simpler for the bulk of the types, which do not use
prepare, I added another common class that does the unsupported
implementation. I also refactored the null driver to reuse the common
classes as well.

BUG=b:278937647
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I4edfa3f2a99d41497581df0374b799fa95755898
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4507333
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>