cryptohome: Make prepare a non-const operation
Unlike most of the driver operations which are "check thing" or "create
thing" the prepare operations are actually supposed to logically modify
the state of the factor and so they really should not be const.

I didn't do this in the initial change that added them to the drivers
because I thought it would be a lot more invasive than it ended up
being, oops.

BUG=b:278937647
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I3a0da3f15745b3f3e8db5651e369d881844c02b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4504744
Reviewed-by: Zi Lin <lziest@google.com>
Reviewed-by: Behnood Momenzadeh <behnoodm@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>