shill: cellular: Add Entitlement check to the anomaly detector
Detect entitlement check failures using the anomaly detector to upload
logs for troubleshooting.
The entitlement check is used only by a few carriers to validate that
hotspot is allowed on the device. The check is only triggered when the
user wants to start a hotspot session, and once a day after that.

BUG=b:280071368
TEST=Trigger an entitlement failure, and verify the crash report is
uploaded.

Change-Id: Idde9aaddc6dc69b6bc17aba9300707a75525d3e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4518035
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Madhav <madhavadas@google.com>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>