cryptohome: Use std::array to implement block_types
The existing code using some template complexity to work around the fact
that arrays in C++ cannot be 0-length. However, it occurred to me that
the stdlib already had to solve this problem with std::array which they
special cased to work with N==0. And span happily works with std::array
arrays instead of the raw kind, so this simplifies things a lot.

BUG=none
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I284fa82d4a354937ae7b3f5e1e8e3447329afcb2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4508114
Reviewed-by: Betul Soysal <betuls@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>