shill: Add new error string for service without device
In the context of b/230401667, it will be helpful to know if we
have any service with CellularRefPtr after the device is
deregistered. Cellular device destructor will not be called
if a cellular device pointer is still present with service.
Currently we see one common message when device
is not associated with service and when device the is disabled.

BUG=b:230401667
TEST=Unit tested on Redrix

Change-Id: I5ccbe1e93120e58a570543e9d587d42e39bb62be
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4508122
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Ujjwal Pande <ujjwalpande@google.com>
Tested-by: Ujjwal Pande <ujjwalpande@google.com>