cryptohome: Relocate the pin-specific stuff
This moves the PIN-specific logic in AuthSession down into the auth
factor driver interface, and the implementation more specifically into
the PIN-specific driver. For the rest of the drivers we just default to
"delays are not currently supported by this factor".

This shouldn't significantly change any of the behavior, other than
maybe reordering things a little or shuffling around the error
locations. It just moves the type-specific logic out of AuthSession.

BUG=b:260904713
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I2f814cd41f63d3859c7abb63ca56bf27cc8cc532
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4510017
Reviewed-by: Behnood Momenzadeh <behnoodm@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>