power: Bluetooth hid increases autosuspend delay
When a Bluetooth hid device is connected, we should increase autosuspend
delay on that device. Users are reporting that devices are lagging after
being left idle for a while and increasing the delay to 2 mins mitigates
this.

This is protected by a Finch controllable flag so that we can measure
the power impact.

BUG=b:278582089
TEST=Verified on Guybrush that connecting keyboard increases delay
     and disconnecting the keyboard puts it back to default

Change-Id: I5b9c1f11b64d31ec92f69d7b93d7d1f3d00a3123
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4509705
Commit-Queue: Puthikorn Voravootivat <puthik@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Auto-Submit: Abhishek Pandit-Subedi <abhishekpandit@google.com>
Tested-by: Abhishek Pandit-Subedi <abhishekpandit@google.com>