diagnostics: Add display_utils class into delegate
Create a class named `display_utils` inside delegate that will be in
charge of all access to drm resources. This is a collection of
functions found in the soon-deprecated `libdrm_utils`,
`display_utils` and `display_fetcher`.

This is part of a bigger effort to migrate all usage of drm and
display related functions into executor-delegate.

BUG=b:281628161
TEST=1.(cr) FEATURES=test  USE="-cros-debug" P2_TEST_FILTER='*' emerge-zork diagnostics

Change-Id: I2ac6c0dac052440bc5edbcf2c0094b290988819f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4515329
Tested-by: Ethan Cheng <yycheng@google.com>
Reviewed-by: Kerker Yang <kerker@chromium.org>
Commit-Queue: Ethan Cheng <yycheng@google.com>