shill: ignore PUK2 lock and allow to use SIM card
The PIN2 is always enabled but it doesn't interfere with connection to
mobile network. To lock SIM with PUK2, an invalid PIN2 needs to be send
three times. It is only possible with command line tools.
SIM card locked with PUK2 still can be used for connection.

BUG=b:278801632
TEST=Run CellularCapability3gppTest

Change-Id: Ie21638c84443198c2a29062c12b10f174e149276
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4513920
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Michał Mazur <michamazur@google.com>
Commit-Queue: Michał Mazur <michamazur@google.com>