hiberman: Simplify metrics file management
Currently clients of the MetricsLogger are supposed to provide
the logger with a file to which to log the metrics. This is clunky
and more so when the client has to to take care of opening and
closing the file when 'hibermeta' is mounted and unmounted.
Instead let the MetricsLogger open the file only when it is about
to flush the metrics buffer.

Also stop having separate files for suspend and a resume metrics
file, which isn't really necessary.

BUG=b:281596038
TEST=hibernate metrics are shown in chrome://histograms

Change-Id: Icee9bece834932a304fb9d56a3838237c2d4f9ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4519674
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>