hiberman: Make MetricsLogger a singleton
Before this change SuspendConductor and ResumeConductor each owned
a MetricsLogger instance, which they passed to other functions that
need to interact with the MetricsLogger. There should be only one
instance of the MetricsLogger and passing it as an argument to
functions is clunky, Make MetricsLogger a singleton instead, which
is more correct semantically and avoids the parameter passing,
though dealing with singleton also has a certain level of clunkiness.

BUG=b:281596038
TEST=hibernate metrics are shown in chrome://histograms

Change-Id: Ibcee74d79707178c1c0d6bdfbd5c8d1c09e3a4eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4519675
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Brian Geffon <bgeffon@chromium.org>