cryptohome: Add an explicit intent check for full auth
This adds an API call to the factor driver interface that works
similarly to the existing IsVerifySupported that checks if an intent is
supported, but for the full auth sequence using Authenticate.

For more consistency this also renames the IsVerifySupported to have
more of a matching name, as the two functions are intended to be doing
the same check "is this intent allowed" check for two different flows.

BUG=b:274782160
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I7fff2fbec786e23e3f0c60669858dd77b9961ef5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4515996
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Reviewed-by: Zi Lin <lziest@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>