cryptohome: Disable using fingerprint for full decrypt
There remains some work to be done to make fingerprint policy more
configurable, before we allow it to be used for full decryption. So for
now, disallow using it for this intent, and only allow it to be used for
verification (unlock).

Note that one significant change this requires is adjusting the
AuthSession behavior to not automatically authorize sessions for
"decrypt" just because they use full auth. Instead, I changed the
behavior to only authorize the session for the subset of intents the
driver allows on full auth.

BUG=b:274782160
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I3edfe0667bb3c0f66038ee28c909fe68c843a703
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4517593
Reviewed-by: Howard Yang <hcyang@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>
Reviewed-by: Betul Soysal <betuls@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>