arc: Add the minimal KeyMint context
To add more functionality to KeyMint, we first need
to make it aware of the KeyMint context, which is
a derived class for AOSP Keymaster context and
mainly defines how Keymaster would be used in
the ChromeOS environment.

This CL creates the pre-requisites required for
ArcKeyMint context. Specifically, it creates Context
Adaptor and Cros Key classes, which will be used further
for ArcKeyMint Context and its integration with KeyMint server.

This CL is mostly based on the Keymaster context in ChromeOS-
platform2/arc/keymaster/context/.

BUG=b:274723555
TEST=emerge-brya-arc-t chromeos-base/arc-keymint

Change-Id: If30238efdd92db41898d5880d40bbfc00ecdf8bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4522703
Reviewed-by: Yao Li <yaohuali@google.com>
Tested-by: Vaibhav Raheja <vraheja@chromium.org>
Commit-Queue: Vaibhav Raheja <vraheja@chromium.org>