libbrillo: modify the root path of the absolute path in unitttest
GetStorageType infers the storage type by matching a specific pattern
from the absolute path of `<root>/sys/block/device/type`. However, in
unittest, the root path is a path to a randomly generated temporary dir,
which might contain the pattern that GetStorageType is looking for.
This CL changes the root path of the absolute path to '/' so that this
issue won't happen.

BUG=b:270086757
TEST=FEATURES=test emerge-nissa libbrillo

Change-Id: I5237d6eaae5712537c0e3b0c1d46aa8e79f8c70c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4515840
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Phoebe Wang <phoebewang@chromium.org>
Tested-by: Phoebe Wang <phoebewang@chromium.org>