shill: do not start the DHCP server if the IPv4 is not ready at upstream
When doing tethering, if the upstream network is not IPv4 ready, then
we should not start the IPv4 DHCP server at the downstream.

Originally, we only check whether the upstream network's ipconfig()
returns valid IPConfig or not. But even if it's not nullptr, it might
be not ready yet. This CL adds the extra check to make sure the IPv4
is ready.

BUG=b:279815763
TEST=emerge-hatch shill

Change-Id: Id6d81b367c50374bf66a9e4215764b7d3b08e83d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4514806
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>