system_api: Adjust API comment to be more accurate
This adjusts the description of the time_available_in field in the
cryptohome auth factor status to be more accurate in the case of
infinite lockout. The current text implies this is left "unset" but
that's not accurate; we actually set the field to the max value.

BUG=none
TEST=comment only

Change-Id: I2978588534793c43c23db7c557763278644eb745
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4518021
Reviewed-by: Howard Yang <hcyang@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Behnood Momenzadeh <behnoodm@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>