libec: Force cmd insize to be 0 for first async BaseCmd
Force the insize of the first async BaseCmd to be zero because the
first async command only schedules the command and does not return any
response with a meaningful size. Restore the insize to its original
value before the execution of the second async command because this is
the command that will return the actual response. Add the proper unit
tests.

BUG=b:116396469
TEST=FEATURES="test" emerge-hatch libec

Change-Id: Ibfab77a1a804c51662264be2e65390938b1176ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4518023
Commit-Queue: Firas Sammoura <fsammoura@google.com>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Reviewed-by: Bobby Casey <bobbycasey@google.com>
Tested-by: Firas Sammoura <fsammoura@google.com>