shill: cellular: Continue APN round robin on Unknown error
There are a few scenarios in which the modems return Unknown error.
Sometimes this error is related to the last APN tried, and using a
different APN will result in a successful connect. Currently, the round
robin does not continue when an Unknown error is returned by
ModemManager. Change that so the round robin also continues when this
error is detected.

BUG=b:269505511
TEST=tast run crota cellular.ShillCellularApn.round_robin_connect_ipv4_default_attach

Change-Id: Icb6b9ad97f616f8b52224b3ad5bbe8f865eb45cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4521722
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>