hiberman: Make unit of duration metrics configurable
Add a unit parameter to the function for reporting duration metrics,
to specify whether the duration should be logged in milliseconds,
seconds, minutes or hours.

While we are at it rename metrics_send_duration_sample() to
log_duration_sample(). The function doesn't send anything
and in the 'metrics' suffix is redundant in the context of
the MetricsLogger.

BUG=b:281596038
TEST=emerge-brya-hibernate hiberman

Change-Id: I3d0f178707e9f90d61fb0456de2803841eb0d022
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4522656
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>