shill: store the Service of the upstream in TetheringManager
Originally, TetheringManager only stores the Network instance of the
upstream network. However, we need to query if the upstream is metered
or not. But the Network class doesn't have the information.

This CL stores the Service of the upstream. With that,
TetheringManager could query if the network is metered with the
Service instance.

BUG=b:275644951
TEST=FEATURES=test emerge-brya shill

Change-Id: I23d82af31ac1502e9af861d28fca87a26db1818c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4517488
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>