diagnostics: Mark audio event as always supported
Though Chromebox may not have a speaker so the event may not be
suitable, this is too rare so we prefer to report it as supported first.
If any OEMs need to develop diagnostic routine on Chromebox devices and
encounter issues, we can see if we need to add a field into Boxster to
tell us the ground truth.

BUG=b:271514913, b:270645724
TEST=cros-health-tool event --category=audio --check_supported
TEST=FEATURES=test emerge-brya diagnostics

Change-Id: Ia8f0c97c955f720acb9a633751920f1f82860ba8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4521361
Commit-Queue: Kerker Yang <kerker@chromium.org>
Tested-by: Kerker Yang <kerker@chromium.org>
Reviewed-by: Denny Huang <dennyh@google.com>