hiberman: Log reason for suspend abort as a metric
Report a metric sample if suspend is aborted due to insufficient
free memory, insufficient free disk space or because the UpdateEngine
was busy. For now report an 'other'/'unknown' error if the creation
of the snapsot image fails for another reason than not having enough
free memory.

BUG=b:232184089
TEST=Platform.Hibernate.Abort metric is shown in chrome://histograms

Change-Id: I7a09ebf629f614b029a7c79f33ce2678a6269ca7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4528425
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>