hiberman: Make sure 'hibermeta' is unmounted when things go wrong
The SuspendConductor and ResumeConductor use error propagation
liberally. When errors occur the 'hibermeta' volume might be
mounted and aborting a function might result in skipping
the corresponding unmount.

This adds the ActiveMount and LogRedirectGuard structs. An
ActiveMount which keeps track of an active mount and unmounts
it when the ActiveMount goes out of scope. Have mount_hibermeta()
and setup_hibermeta_lv() return ActiveMounts. In some cases the
function that mounts 'hibermeta' doesn't unmount it itself,
instead the unmount is done in a sub-function. In these cases
the ActiveMount is passed to the sub-function (by ownership).
This is a bit clunky, but fortunately there are only a couple
of instances. LogRedirectGuard is similar to ActiveMount, it
keeps track of an open log file, which must be closed before
'hibermeta' can be unmounted. When the LogRedirectGuard goes
out of scope it redirects the hibernate log to a memory
buffer, which results in closing the log file.

It is important that the log file is closed before 'hibermeta'
is unmounted. In Rust local variables within the same scope
are dropped in reverse declaration order, so the
LogRedirectGuard must be declared after the corresponding
ActiveMount, which coincides with the natural program flow. We
pass both an ActiveMount and a LogRedirectGuard as parameters
to suspend_system(). Function parameters are also dropped in
reverse order, so it's important that the ActiveMount is the
first parameter and the LogRedirectGuard the second.

BUG=b:282161297
TEST=hibermeta is unmounted when errors occur

Change-Id: Ic0b1c564ea75cef9fbed822c7996aed98abfa0c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4526615
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>